requirements.txt
pykakasi == 2.1.1
python-dotenv == 0.17.1
firebase-admin == 5.0.1